repl_test: rename module wasm_interp->wasm
