+++ crates/repl_test/src/wasm.rs
